Count concordant reads

+++ src/main.rs
@@ -354,7 +354,7 @@ fn bam_stats(path: String, end_margin: usize) {
             total_read_length += record.query_len() as u64 + record.cigar().hard_clipping(true) as u64 + record.cigar().hard_clipping(false) as u64;
             total_primary_aligned_read_length += (record.aligned_query_end() - record.aligned_query_start()) as u64;
             primary_alignment += 1;
-            if record.aligned_query_start() <= end_margin && (record.query_len() as u64 + record.cigar().hard_clipping(true) as u64 + record.cigar().hard_clipping(false) as u64 - record.aligned_query_end()) <= end_margin {
+            if record.aligned_query_start() as usize <= end_margin && (record.query_len() as usize + record.cigar().hard_clipping(true) as usize + record.cigar().hard_clipping(false) as usize - record.aligned_query_end() as usize) <= end_margin {
                 concordant_reads += 1;
             }
         } else if !record.flag().is_mapped()
